refactor: shift layout to dashboard view
