+++ pages/profile/api-key.tsx
@@ -3,8 +3,8 @@ import React, { ReactElement } from 'react';
 import UserLayout from '../../src/features/common/Layout/UserLayout/UserLayout';
 import { useTranslation } from 'next-i18next';
 
-import ApiKey from '../../src/features/user/Settings/ApiKey';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import ApiKey from '../../src/features/user/Settings/ApiKey';
 
 function EditProfilePage(): ReactElement {
   const { t } = useTranslation('me');
