+++ src/features/user/Settings/ApiKey/ApiKey.module.scss
@@ -1,4 +1,4 @@
-@import './../../../theme/theme';
+@import '../../../../theme/theme';
 
 .apiPage {
   display: flex;
